+++ ChangeLog.md
@@ -1,4 +1,4 @@
-﻿## 08/21/2018 V7.0.0
+﻿## 08/26/2018 V7.0.0
 
 - Draw(SpriteBatch batch, Point position, Point size, Font font) has been removed.
 - Cell/CellState have a Decorators list which are used to add extra glyph draws to individual cells.
@@ -18,8 +18,7 @@
 - Surface.RenderArea changed to Surface.ViewPort
 - Readded Zone and HotSpot types.
 - Removed random level generation.
-- Added Entities.EntityManager which helps control visibility and offsets based on a parent console.
-
+- Added Entities.EntityManager which helps control entity visibility and offsets based on a parent console. Also handles zones/hotspots.
 
 ## 06/11/2018